added topk_connect file
