+++ torch_geometric/nn/pool/connect/__init__.py
@@ -1,5 +1,7 @@
 from .base import Connect
+from .topk_connect import TopkConnect
 
 __all__ = [
     'Connect',
+    'TopkConnect',
 ]
