+++ torch_geometric/nn/pool/topk_pool.py
@@ -5,10 +5,9 @@ from torch import Tensor
 from torch.nn import Parameter
 
 from torch_geometric.nn.inits import uniform
-from torch_geometric.nn.pool.connect import Connect
-from torch_geometric.nn.pool.select import SelectOutput, TopkSelect
+from torch_geometric.nn.pool.connect import TopkConnect
+from torch_geometric.nn.pool.select import TopkSelect
 from torch_geometric.utils import softmax
-from torch_geometric.utils.num_nodes import maybe_num_nodes
 
 
 class TopkReduce(torch.nn.Module):
@@ -20,41 +19,6 @@ class TopkReduce(torch.nn.Module):
         x = x[perm] * weight[perm].view(-1, 1)
         x = self.multiplier * x if self.multiplier != 1 else x
         return x
-
-
-class TopkConnect(Connect):
-    def forward(
-            cluster: SelectOutput, edge_index: Tensor,
-            edge_attr: Optional[Tensor]) -> Tuple[Tensor, Optional[Tensor]]:
-        return filter_adj(
-            edge_index,
-            edge_attr,
-            cluster.node_index,
-            cluster.num_nodes,
-        )
-
-
-def filter_adj(
-    edge_index: Tensor,
-    edge_attr: Optional[Tensor],
-    perm: Tensor,
-    num_nodes: Optional[int] = None,
-) -> Tuple[Tensor, Optional[Tensor]]:
-    num_nodes = maybe_num_nodes(edge_index, num_nodes)
-
-    mask = perm.new_full((num_nodes, ), -1)
-    i = torch.arange(perm.size(0), dtype=torch.long, device=perm.device)
-    mask[perm] = i
-
-    row, col = edge_index[0], edge_index[1]
-    row, col = mask[row], mask[col]
-    mask = (row >= 0) & (col >= 0)
-    row, col = row[mask], col[mask]
-
-    if edge_attr is not None:
-        edge_attr = edge_attr[mask]
-
-    return torch.stack([row, col], dim=0), edge_attr
 
 
 class TopKPooling(torch.nn.Module):
